t6

+++ animrec/videoin.py
@@ -43,7 +43,7 @@ class VideoIn:
             width = almau.closenr(resolution[0],32)
             height = almau.closenr(resolution[1],16)
             self.resolution = (width,height)
-            self.resolution = (zoom[2]*self.resolution[0], zoom[3]*self.resolution[1])
+            self.resolution = (int(zoom[2]*self.resolution[0]), int(zoom[3]*self.resolution[1]))
             self.camera.resolution = self.resolution
             self.camera.framerate = framerate
             self.camera.zoom = zoom
